Move all database helpers from vote.py to vote_session.py
